优化

+++ sod/sod-dm/sod-dm-rpc-service/src/main/java/com/piesat/dm/rpc/service/special/DatabaseSpecialServiceImpl.java
@@ -202,9 +202,12 @@ public class DatabaseSpecialServiceImpl extends BaseService<DatabaseSpecialEntit
     @Transactional
     @Override
     public void deleteById(String id) {
-        this.databaseSpecialAuthorityDao.deleteBySdbId(id);
-        this.databaseSpecialReadWriteDao.deleteBySdbId(id);
-        this.delete(id);
+        if (StringUtils.isNotEmpty(id)){
+            this.databaseSpecialAuthorityDao.deleteBySdbId(id);
+            this.databaseSpecialReadWriteDao.deleteBySdbId(id);
+            databaseDao.deleteByTdbId(id);
+            this.delete(id);
+        }
     }
 
     @Transactional
